+++ ft_putpixel.c
@@ -6,7 +6,7 @@
 /*   By: wvan-dam <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/05/16 13:57:21 by wvan-dam       #+#    #+#                */
-/*   Updated: 2019/06/07 16:36:33 by pholster      ########   odam.nl         */
+/*   Updated: 2019/06/07 19:04:25 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ unsigned int			*palettemaker(t_ftl *ftl)
 }
 
 static unsigned int			calc_interpolate(unsigned int color1,
-	unsigned int color2, double frac)
+	unsigned int color2, float frac)
 {
 	unsigned int		result;
 
@@ -55,7 +55,7 @@ static unsigned int			calc_interpolate(unsigned int color1,
 }
 
 static unsigned int			interpolate(unsigned int color1,
-	unsigned int color2, double frac)
+	unsigned int color2, float frac)
 {
 	unsigned int		result;
 	unsigned int		rgb1;
